+++ driven/storage/database.go
@@ -136,6 +136,11 @@ func (m *database) start() error {
 		return err
 	}
 
+	err = m.ApplyGroupsAttributesTransition(client, groups)
+	if err != nil {
+		return err
+	}
+
 	//asign the db, db client and the collections
 	m.db = db
 	m.dbClient = client
@@ -195,11 +200,9 @@ func (m *database) applyUsersChecks(users *collectionWrapper) error {
 
 	indexes, _ := users.ListIndexes()
 	indexMapping := map[string]interface{}{}
-	if indexes != nil {
-		for _, index := range indexes {
-			name := index["name"].(string)
-			indexMapping[name] = index
-		}
+	for _, index := range indexes {
+		name := index["name"].(string)
+		indexMapping[name] = index
 	}
 
 	if indexMapping["external_id_1"] == nil {
@@ -249,11 +252,10 @@ func (m *database) applyGroupsChecks(groups *collectionWrapper) error {
 
 	indexes, _ := groups.ListIndexes()
 	indexMapping := map[string]interface{}{}
-	if indexes != nil {
-		for _, index := range indexes {
-			name := index["name"].(string)
-			indexMapping[name] = index
-		}
+
+	for _, index := range indexes {
+		name := index["name"].(string)
+		indexMapping[name] = index
 	}
 
 	if indexMapping["client_id_1"] == nil {
@@ -482,11 +484,9 @@ func (m *database) applyEventsChecks(events *collectionWrapper) error {
 
 	indexes, _ := events.ListIndexes()
 	indexMapping := map[string]interface{}{}
-	if indexes != nil {
-		for _, index := range indexes {
-			name := index["name"].(string)
-			indexMapping[name] = index
-		}
+	for _, index := range indexes {
+		name := index["name"].(string)
+		indexMapping[name] = index
 	}
 
 	if indexMapping["event_id_1_group_id_1_client_id_1"] == nil {
@@ -589,13 +589,12 @@ func (m *database) applyPostsChecks(posts *collectionWrapper) error {
 
 	indexes, _ := posts.ListIndexes()
 	indexMapping := map[string]interface{}{}
-	if indexes != nil {
 
-		for _, index := range indexes {
-			name := index["name"].(string)
-			indexMapping[name] = index
-		}
+	for _, index := range indexes {
+		name := index["name"].(string)
+		indexMapping[name] = index
 	}
+
 	if indexMapping["client_id_1"] == nil {
 		err := posts.AddIndex(
 			bson.D{
@@ -801,7 +800,7 @@ func (m *database) ApplyMembershipTransition(client *mongo.Client, groups *colle
 
 	var migrationGroup []model.Group
 	err := groups.Find(bson.D{
-		{"members.id", bson.M{"$exists": true}},
+		{Key: "members.id", Value: bson.M{"$exists": true}},
 	}, &migrationGroup, nil)
 	if err != nil {
 		return err
@@ -816,8 +815,8 @@ func (m *database) ApplyMembershipTransition(client *mongo.Client, groups *colle
 			}
 
 			_, err = groups.UpdateManyWithContext(sessionContext, bson.D{}, bson.D{
-				{"$set", bson.D{
-					{"stats", model.GroupStats{}},
+				{Key: "$set", Value: bson.D{
+					{Key: "stats", Value: model.GroupStats{}},
 				}},
 			}, nil)
 			if err != nil {
@@ -856,12 +855,12 @@ func (m *database) ApplyMembershipTransition(client *mongo.Client, groups *colle
 				}
 
 				_, err = groups.UpdateOneWithContext(sessionContext, bson.D{
-					{"client_id", group.ClientID},
-					{"_id", group.ID},
+					{Key: "client_id", Value: group.ClientID},
+					{Key: "_id", Value: group.ID},
 				}, bson.D{
-					{"$set", bson.D{
-						{"members", nil},
-						{"stats", stats},
+					{Key: "$set", Value: bson.D{
+						{Key: "members", Value: nil},
+						{Key: "stats", Value: stats},
 					}},
 				}, nil)
 				if err != nil {
@@ -902,7 +901,7 @@ func (m *database) ApplyDefaultGroupSettings(client *mongo.Client, groups *colle
 		var migrationGroup []model.Group
 
 		filter := bson.D{
-			{"settings", bson.M{"$exists": false}},
+			{Key: "settings", Value: bson.M{"$exists": false}},
 		}
 
 		err = groups.FindWithContext(sessionContext, filter, &migrationGroup, nil)
@@ -912,8 +911,8 @@ func (m *database) ApplyDefaultGroupSettings(client *mongo.Client, groups *colle
 
 		if len(migrationGroup) > 0 {
 			_, err = groups.UpdateManyWithContext(sessionContext, filter, bson.D{
-				{"$set", bson.D{
-					{"settings", model.DefaultGroupSettings()},
+				{Key: "$set", Value: bson.D{
+					{Key: "settings", Value: model.DefaultGroupSettings()},
 				}},
 			}, nil)
 			if err != nil {
@@ -936,6 +935,47 @@ func (m *database) ApplyDefaultGroupSettings(client *mongo.Client, groups *colle
 	}
 
 	log.Println("group settings migration passed")
+	return nil
+}
+
+func (m *database) ApplyGroupsAttributesTransition(client *mongo.Client, groups *collectionWrapper) error {
+	log.Println("apply group attributes migration.....")
+
+	err := client.UseSession(context.Background(), func(sessionContext mongo.SessionContext) error {
+		err := sessionContext.StartTransaction()
+		if err != nil {
+			log.Printf("error starting a transaction - %s", err)
+			return err
+		}
+
+		filter := bson.D{
+			{Key: "attributes", Value: bson.M{"$exists": false}},
+		}
+		_, err = groups.UpdateManyWithContext(sessionContext, filter, bson.D{
+			{Key: "$set", Value: bson.D{
+				{Key: "attributes.category", Value: "$category"},
+				{Key: "attributes.tags", Value: "$tags"},
+			}},
+		}, nil)
+		if err != nil {
+			abortTransaction(sessionContext)
+			return err
+		}
+
+		//commit the transaction
+		err = sessionContext.CommitTransaction(sessionContext)
+		if err != nil {
+			fmt.Println(err)
+			return err
+		}
+		return nil
+	})
+
+	if err != nil {
+		return err
+	}
+
+	log.Println("group attributes migration passed")
 	return nil
 }